+++ enum/src/main/java/org/grnet/cat/enums/ValidationStatus.java
@@ -4,6 +4,5 @@ public enum ValidationStatus {
 
     PENDING,
     REVIEW,
-    APPROVED,
-    REJECTED
+    APPROVED
 }